fix(ci): resolve all CI failures after workflow fix [ESCALATION] (#34)
* fix(ci): resolve all CI failures after workflow fix [ESCALATION]

Four fixes to unblock CI after the secrets-context patch (PR #33):

1. Remove invalid secrets context from step-level if in doppler-verify
   (supersedes PR #33 — 'secrets' not allowed in step if expressions).

2. Add 'Build shared package' step before 'Lint' in the lint job.
   ESLint strictTypeChecked needs @estalara/shared dist to resolve
   imported types; without the build it reports 23 no-unsafe-* errors.

3. Add permissions: pull-requests: read to gitleaks-scan job.
   gitleaks-action@v2 needs this permission on pull_request events
   to post review comments; without it the action exits 403.

4. Add apps/auto-detect/src/test_smoke.py with one test that calls
   is_ready(). pytest src/ exited with code 5 (no tests collected)
   because auto-detect had no test file under src/. One collected
   test resolves the exit-5 failure.

5. Drop --user flag from ClickHouse curl calls when CH_PASS is empty.
   Docker ClickHouse default user requires no credentials; sending
   Basic auth with an empty password yields a 401 on recent images.
   Fixed in both migrate.sh and smoke-test.sh.


* fix(ci): set CLICKHOUSE credentials on service container [ESCALATION]

Recent clickhouse/clickhouse-server images disable network access for
the default user when neither CLICKHOUSE_USER nor CLICKHOUSE_PASSWORD
is set as a container env var. This caused curl: (22) 401 despite the
correct HTTP endpoint. Setting CLICKHOUSE_USER=default and
CLICKHOUSE_PASSWORD="" (empty string) re-enables network access and
lets the no-auth curl calls in migrate.sh / smoke-test.sh succeed.


* fix(ci): use non-empty CLICKHOUSE_PASSWORD for docker service [ESCALATION]

GitHub Actions passes an empty YAML string as -e "CLICKHOUSE_PASSWORD"
(no value) to docker create, which the container sees as unset. The
ClickHouse entrypoint only enables network access when CLICKHOUSE_PASSWORD
is a non-empty env var. Set password to "clickhouse" on both the service
container and the migrate/smoke-test steps so the scripts auth correctly.


* fix(ci): restore auth in migrate.sh _apply_file LOCAL=1 path [ESCALATION]

The LOCAL=1 branch of _apply_file lost its credential handling when we
removed the unconditional --user flag. Now that the ClickHouse service
container requires a password (CLICKHOUSE_PASSWORD=clickhouse), the
no-auth curl was yielding 401. Unified the auth logic: both LOCAL and
non-LOCAL paths now check CH_PASS and pass -u user:pass when set.


* fix(ci): execute each migration statement individually [ESCALATION]

ClickHouse HTTP processes one query per request. Migration 0002 contains
two statements (CREATE TABLE + CREATE MATERIALIZED VIEW) which caused a
400 when sent together. Refactor _apply_file to split on statement
boundaries and execute each statement via a dedicated _ch_send helper.


* fix(ci): show ClickHouse error body on HTTP failure [ESCALATION]

Replace -sSf with -sS --fail-with-body so CI logs display the actual
ClickHouse error message when a migration returns 400, instead of only
reporting the exit code. Needed to diagnose the remaining smoke failure.


* fix(data): use UInt64 for page_count SimpleAggregateFunction [ESCALATION]

ClickHouse 26.3 (official build) validates that SimpleAggregateFunction
column type matches the aggregate function return type. sum(UInt32)
returns UInt64, so the column must be UInt64. The MV SELECT already
uses countIf which returns UInt64, so no change needed there.


---------

+++ infra/clickhouse/migrations/0002_create_session_summary_mv.sql
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS session_summary
 
     -- Aggregated session features
     ended_at         SimpleAggregateFunction(max,     DateTime64(3, 'UTC')),
-    page_count       SimpleAggregateFunction(sum,     UInt32),
+    page_count       SimpleAggregateFunction(sum,     UInt64),
     listing_ids_seen AggregateFunction(uniq,          String),
     has_chat         SimpleAggregateFunction(max,     UInt8),
     has_inquiry      SimpleAggregateFunction(max,     UInt8),
